requirements.txt
google-generativeai>=0.3.0
pathlib
python-dotenv>=1.0.0 